fix previous-release bug and MY Y-axis to 0.6500
Fix a bug where if two opposing directions are held and the one that's
not in effect is released, that axis is reset to neutral.

MY now modifies the Y-axis to 0.6500 instead (there doesn't seem to be
able point in pointing in Y-smash territory since unmodified already
does this).

Use constants with more readable names (rectangular co-ordinates inside
the unit circle) that map to the raw output value (between -80 to 80).