add jump_delay_ms argument for delaying X press
Add an argument which delays the X button's press when the corresponding
key is pressed. This makes the timing of short hops more lenient, at the
cost of some input delay. It is difficult to release the X button fast
enough to short hop within the 2-frame window for characters with a
3-frame jumpsquat with mechanical switches with the typical 2mm
actuation and 4mm travel.

Note that the tournament legality of enabling this jump delay feature is
questionable, as it can be considered a macro of a kind since it results
in the jump being input on "future frames".